Backend/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1
djangorestframework==3.13.1
fire==0.4.0
fonttools==4.34.4
lxml==4.9.1
numpy==1.23.1
opencv-python==4.6.0.66
pdf2docx==0.5.5
PyMuPDF==1.20.1
python-docx==0.8.11
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
termcolor==1.1.0
